+++ src/Files.App/Views/Properties/SecurityPage.xaml
@@ -176,7 +176,8 @@
 						Padding="8,6,4,6"
 						VerticalAlignment="Center"
 						AutomationProperties.Name="{helpers:ResourceString Name=AdvancedPermissions}"
-						Command="{x:Bind OpenSecurityAdvancedPageCommand, Mode=OneWay}"
+						Command="{x:Bind SecurityViewModel.OpenSecurityAdvancedPageCommand, Mode=OneWay}"
+						CommandParameter="{x:Bind Frame, Mode=OneWay}"
 						ToolTipService.ToolTip="{helpers:ResourceString Name=AdvancedPermissions}">
 						<StackPanel Orientation="Horizontal" Spacing="4">
 							<TextBlock Text="{helpers:ResourceString Name=AdvancedPermissions}" />
